# openapi.yml
openapi: 3.0.3
info:
  title: Corazón Migrante API
  version: 1.0.0
  description: "API pública y administrativa de Corazón Migrante.\n\nPrefijo global: `/api/v1`.\n\nContrato de respuesta:\n\
    - Respuesta simple: `{ data, meta }`.\n- Respuesta paginada: `{ data, pagination, meta }`.\n- Error: `{ error: { code,\
    \ message, details }, meta }`.\n\nSeguridad:\n- Los endpoints públicos están marcados como `public` en su descripción.\n\
    - Los endpoints privados usan `Bearer JWT`.\n- Algunos endpoints administrativos requieren permisos específicos además\
    \ del rol.\n\nDisponibilidad de terapeutas:\n- `GET /booking/availability` calcula los horarios disponibles por terapeuta.\n\
    - Parte de los horarios recurrentes configurados en `therapist_schedules`.\n- Resta bloqueos activos de agenda (`therapist_blocked_times`).\n\
    - Resta citas activas/ocupadas (`appointments` con status `REQUESTED` o `CONFIRMED`).\n- El rango máximo soportado es\
    \ 31 días.\n\nLímites de uso (rate limiting):\n- Límite global por IP: 120 solicitudes / 60s. Al excederlo, la API responde\
    \ `429`\n  con `{ error: { code: \"HTTP_429\", ... } }`.\n- `POST /auth/login`, `/auth/register/*`: 10 solicitudes / 60s\
    \ por IP.\n- `POST /auth/password-reset/request` y `/confirm`: 5 solicitudes / 60s por IP.\n- El cuerpo de las solicitudes\
    \ JSON está limitado a 2 MB (`413` si se excede).\n"
servers:
- url: https://api.corazondemigrante.com/api/v1
  description: Producción
...
security:
- bearerAuth: []
paths:
...
  /publications/categories:
    get:
      tags:
      - Publications public
      summary: Listar categorías públicas de contenido
      security: []
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
  /publications/tags:
    get:
      tags:
      - Publications public
      summary: Listar tags públicos de contenido
      security: []
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
  /admin/content/publications:
    get:
      tags:
      - Content admin
      summary: Listar publicaciones administrativas
      description: Requiere `content:read`. Permite filtrar por estado, tipo editorial, acceso y búsqueda.
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PPage'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/PLimit'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Order'
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Q'
      - name: status
        in: query
        schema:
          type: string
          enum:
          - DRAFT
          - IN_REVIEW
          - SCHEDULED
          - PUBLISHED
          - ARCHIVED
      - name: publicationType
        in: query
        schema:
          type: string
          enum:
          - NEWS
          - COLUMN
          - OPINION
          - INTERVIEW
          - REPORT
          - ANALYSIS
      - name: accessType
        in: query
        schema:
          type: string
          enum:
          - PUBLIC
          - PREMIUM
          - INTERNAL_ONLY
      - name: categorySlug
        in: query
        schema:
          type: string
      - name: tagSlug
        in: query
        schema:
          type: string
      - name: pageSlug
        in: query
        schema:
          type: string
        description: Slug de página pública configurada en seoMetadata.embedPageSlugs.
      - name: authorId
        in: query
        schema:
          type: string
          format: uuid
      responses:
        '200':
          $ref: '#/components/responses/PaginatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
      - Content admin
      summary: Crear publicación
      description: Requiere `content:write`. Permite crear contenido público, premium o interno.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePublicationRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/content/publications/{id}:
    get:
      tags:
      - Content admin
      summary: Obtener publicación administrativa
      parameters:
      - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags:
      - Content admin
      summary: Actualizar publicación
      parameters:
      - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePublicationRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/content/publications/{id}/publish:
    post:
      tags:
      - Content admin
      summary: Publicar publicación
      parameters:
      - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/content/publications/{id}/schedule:
    post:
      tags:
      - Content admin
      summary: Programar publicación
      parameters:
      - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - scheduledAt
              properties:
                scheduledAt:
                  type: string
                  format: date-time
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/content/publications/{id}/archive:
    post:
      tags:
      - Content admin
      summary: Archivar publicación
      parameters:
      - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/content/categories:
    get:
      tags:
      - Content admin
      summary: Listar categorías de contenido
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
      - Content admin
      summary: Crear categoría de contenido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentCategoryRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/content/categories/{id}:
    patch:
      tags:
      - Content admin
      summary: Actualizar categoría de contenido
      parameters:
      - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContentCategoryRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/content/tags:
    get:
      tags:
      - Content admin
      summary: Listar tags de contenido
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
      - Content admin
      summary: Crear tag de contenido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentTagRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/content/authors:
    get:
      tags:
      - Content admin
      summary: Listar autores de contenido
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
      - Content admin
      summary: Crear autor de contenido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentAuthorRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/content/authors/{id}:
    patch:
      tags:
      - Content admin
      summary: Actualizar autor de contenido
      parameters:
      - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContentAuthorRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/content/subscribers:
    get:
      tags:
      - Content admin
      summary: Listar suscriptores premium vinculados a pacientes
      description: Requiere `content:read`. Un suscriptor debe estar asociado a una cuenta de paciente existente.
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PPage'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/PLimit'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Order'
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Q'
      - name: status
        in: query
        schema:
          type: string
          enum:
          - ACTIVE
          - UNSUBSCRIBED
          - SUSPENDED
      responses:
        '200':
          $ref: '#/components/responses/PaginatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
      - Content admin
      summary: Crear o actualizar suscriptor premium paciente
      description: Requiere `content:write`. `userId` debe corresponder a un usuario con perfil de paciente y el email debe
        coincidir con ese usuario.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertContentSubscriberRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/content/subscribers/{userId}/subscription:
    patch:
      tags:
      - Content admin
      summary: Actualizar suscripción premium por usuario paciente
      description: Requiere `content:write`. Alias estable para actualizar o crear suscripción usando `userId` del paciente.
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContentSubscriberRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/content/subscribers/{id}:
    patch:
      tags:
      - Content admin
      summary: Actualizar suscriptor premium paciente
      description: Requiere `content:write`.
      parameters:
      - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContentSubscriberRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Página actual. Tiene prioridad sobre `p_page`.
    PPage:
      name: p_page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
      description: Alias legacy aceptado para paginación. `page` tiene prioridad cuando ambos llegan.
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Tamaño de página histórico.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Tamaño de página recomendado para frontend/smoke. Tiene prioridad sobre `pageSize` y `p_limit`.
    PLimit:
      name: p_limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Alias legacy aceptado para tamaño de página. `limit` y `pageSize` tienen prioridad cuando llegan juntos.
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
      description: Búsqueda textual canónica.
    Q:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: Alias corto para búsqueda. `search` tiene prioridad cuando ambos llegan.
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: createdAt
      description: Campo de orden. El backend normaliza alias camelCase/snake_case seguros y evita ORDER BY con columnas inexistentes.
...
    Order:
      name: order
      in: query
      required: false
      schema:
        type: string
        enum:
        - asc
        - desc
        default: desc
...
  responses:
    OkEnvelope:
      description: Respuesta exitosa envuelta en `{ data, meta }`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeObject'
    CreatedEnvelope:
      description: Recurso creado. Respuesta envuelta en `{ data, meta }`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeObject'
    PaginatedEnvelope:
      description: Lista paginada envuelta en `{ data, pagination, meta }`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PaginatedEnvelope'
    ValidationError:
      description: Error de validación o contrato de entrada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: HTTP_400
              message:
              - property debug should not exist
              details:
              - property debug should not exist
            meta:
              requestId: 315c2096-df71-447a-92ad-2f9329316896
              timestamp: '2026-07-05T15:13:13.536Z'
...
    NotFoundError:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ConflictError:
      description: Conflicto de unicidad, solapamiento o regla de negocio.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Meta:
      type: object
      required:
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    EnvelopeObject:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          oneOf:
          - type: object
            nullable: true
          - type: array
            items: {}
          - type: string
          - type: boolean
          - type: number
        meta:
          $ref: '#/components/schemas/Meta'
    Pagination:
      type: object
      required:
      - page
      - pageSize
      - limit
      - total
      - totalPages
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 138
        totalPages:
          type: integer
          example: 7
    PaginatedEnvelope:
      type: object
      required:
      - data
      - pagination
      - meta
      properties:
        data:
          type: array
          items: {}
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorEnvelope:
      type: object
      required:
      - error
      - meta
      properties:
        error:
          type: object
          required:
          - code
          - message
          - details
          properties:
            code:
              type: string
              example: HTTP_400
            message:
              oneOf:
              - type: string
              - type: array
                items:
                  type: string
            details:
              type: array
              items: {}
        meta:
          $ref: '#/components/schemas/Meta'
...
    CreatePublicationRequest:
      type: object
      required:
      - authorId
      - categoryId
      - title
      - summary
      - body
      properties:
        authorId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        coverFileId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 4
          maxLength: 220
        slug:
          type: string
          minLength: 4
          maxLength: 240
          description: Si se omite, el backend lo genera desde el título.
        summary:
          type: string
          minLength: 10
          maxLength: 1000
        body:
          type: string
          minLength: 20
          maxLength: 50000
        audioTranscript:
          type: string
        publicationType:
          type: string
          enum:
          - NEWS
          - COLUMN
          - OPINION
          - INTERVIEW
          - REPORT
          - ANALYSIS
          default: NEWS
        accessType:
          type: string
          enum:
          - PUBLIC
          - PREMIUM
          - INTERNAL_ONLY
          default: PUBLIC
        scheduledAt:
          type: string
          format: date-time
        commentsEnabled:
          type: boolean
          default: true
        reactionsEnabled:
          type: boolean
          default: true
        tagIds:
          type: array
          maxItems: 20
          items:
            type: string
            format: uuid
        seoMetadata:
          type: object
          additionalProperties: true
    UpdatePublicationRequest:
      type: object
      properties:
        authorId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        coverFileId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 4
          maxLength: 220
        slug:
          type: string
          minLength: 4
          maxLength: 240
        summary:
          type: string
          minLength: 10
          maxLength: 1000
        body:
          type: string
          minLength: 20
          maxLength: 50000
        audioTranscript:
          type: string
        publicationType:
          type: string
          enum:
          - NEWS
          - COLUMN
          - OPINION
          - INTERVIEW
          - REPORT
          - ANALYSIS
        accessType:
          type: string
          enum:
          - PUBLIC
          - PREMIUM
          - INTERNAL_ONLY
        scheduledAt:
          type: string
          format: date-time
        commentsEnabled:
          type: boolean
        reactionsEnabled:
          type: boolean
        tagIds:
          type: array
          maxItems: 20
          items:
            type: string
            format: uuid
        seoMetadata:
          type: object
          additionalProperties: true
    CreateContentCategoryRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        isActive:
          type: boolean
          default: true
        sortOrder:
          type: integer
          minimum: 0
          maximum: 10000
    UpdateContentCategoryRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        slug:
          type: string
          minLength: 2
          maxLength: 140
        description:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
          minimum: 0
          maximum: 10000
    CreateContentTagRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        slug:
          type: string
          minLength: 2
          maxLength: 100
    CreateContentAuthorRequest:
      type: object
      required:
      - displayName
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 180
        userId:
          type: string
          format: uuid
        headline:
          type: string
          minLength: 2
          maxLength: 220
        bio:
          type: string
        avatarFileId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
        metadata:
          type: object
          additionalProperties: true
    UpdateContentAuthorRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 180
        userId:
          type: string
          format: uuid
        headline:
          type: string
          minLength: 2
          maxLength: 220
        bio:
          type: string
        avatarFileId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
        metadata:
          type: object
          additionalProperties: true
...
    UpsertContentSubscriberRequest:
      type: object
      required:
      - userId
      - email
      properties:
        userId:
          type: string
          format: uuid
          description: Usuario que debe tener perfil de paciente.
        email:
          type: string
          format: email
          description: Debe coincidir con el correo del usuario paciente seleccionado.
        displayName:
          type: string
          minLength: 2
          maxLength: 180
        status:
          type: string
          enum:
          - ACTIVE
          - UNSUBSCRIBED
          - SUSPENDED
          default: ACTIVE
        subscriptionTier:
          type: string
          enum:
          - FREE
          - PREMIUM
          default: FREE
        premiumUntil:
          type: string
          format: date-time
          nullable: true
        source:
          type: string
          example: admin
        consentMetadata:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
    UpdateContentSubscriberRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: Usuario que debe tener perfil de paciente.
        email:
          type: string
          format: email
          description: Debe coincidir con el correo del usuario paciente seleccionado.
        displayName:
          type: string
          minLength: 2
          maxLength: 180
        status:
          type: string
          enum:
          - ACTIVE
          - UNSUBSCRIBED
          - SUSPENDED
        subscriptionTier:
          type: string
          enum:
          - FREE
          - PREMIUM
        premiumUntil:
          type: string
          format: date-time
          nullable: true
        source:
          type: string
        consentMetadata:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
...